Address refresh path helper review comments
